+++ hypervisor/common/trusty_hypercall.c
@@ -101,7 +101,7 @@ int32_t hcall_initialize_trusty(struct acrn_vcpu *vcpu, uint64_t param)
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_save_restore_sworld_ctx(struct acrn_vcpu *vcpu)
+int32_t hcall_save_restore_sworld_ctx(struct acrn_vcpu *vcpu)
 {
 	struct acrn_vm *vm = vcpu->vm;
